Unify aggregate catalog build behind a typed registry loop.
Extract catalog emission to aggregate_catalog_build.py with one entry-driven loop and Callable config caps instead of getattr on string field names.
